+++ CHANGELOG.md
@@ -4,6 +4,8 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/), and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Expose task-specific compute resources in `run_parallel`/`build_parallel_task_graph` via a `per_task_resources` callable (requires `flat_node_args`).
 
 ## [1.0.27] - 2025-07-10
 ### Added